+++ scripts/api.js
@@ -5,14 +5,6 @@ import { ItemPileInventory } from "./formapplications/itemPileInventory.js";
 import DropItemDialog from "./formapplications/dropItemDialog.js";
 import HOOKS from "./hooks.js";
 import { hotkeyState } from "./hotkeys.js";
-import {
-    getActorCurrencies,
-    getActorItems,
-    getItemPileItemsForActor,
-    getItemPileName,
-    getItemPileTokenImage,
-    getItemPileTokenScale
-} from "./lib/lib.js";
 
 export default class API {
 
@@ -157,7 +149,7 @@ export default class API {
             })
         }
 
-        return itemPileSocket.executeAsGM(SOCKET_HANDLERS.CREATE_PILE, position, { items, pileActorName });
+        return itemPileSocket.executeAsGM(SOCKET_HANDLERS.CREATE_PILE, position, { pileActorName, items });
     }
 
     /**
@@ -821,6 +813,14 @@ export default class API {
         return ItemPileInventory.rerenderActiveApp(inPileUuid, deleted);
     }
 
+    /**
+     * Splits an item pile's content between all players (or a specified set of target actors).
+     *
+     * @param itemPile {Token/TokenDocument/Actor}                                              The item pile to split
+     * @param targets {boolean/TokenDocument/Actor/Array<TokenDocument/Actor>} [targets=false]  The targets to receive the split contents
+     * @param instigator {boolean/TokenDocument/Actor} [instigator=false]                       Whether this was triggered by a specific actor
+     * @returns {Promise<object>}
+     */
     static async splitItemPileContents(itemPile, { targets = false, instigator = false }={}){
 
         if (!lib.isValidItemPile(itemPile)) return false;
@@ -847,6 +847,9 @@ export default class API {
         }
 
         const actorUuids = (targets || lib.getPlayersForItemPile(itemPileActor).map(u => u.character)).map(actor => lib.getUuid(actor));
+
+        const hookResult = Hooks.call(HOOKS.PILE.PRE_SPLIT_INVENTORY, itemPile, targets, game.user.id, instigator);
+        if (hookResult === false) return;
 
         return itemPileSocket.executeAsGM(SOCKET_HANDLERS.SPLIT_PILE, itemPileUuid, actorUuids, game.user.id, instigator);
 
@@ -895,7 +898,7 @@ export default class API {
         await itemPileSocket.executeForEveryone(
             SOCKET_HANDLERS.CALL_HOOK,
             HOOKS.PILE.SPLIT_INVENTORY,
-            itemPile,
+            itemPileUuid,
             transferData,
             userId,
             instigator
@@ -909,6 +912,7 @@ export default class API {
             userId: userId,
             instigator: instigator
         };
+
         await API._executeItemPileMacro(itemPileUuid, macroData);
 
         const shouldBeDeleted = await API._checkItemPileShouldBeDeleted(itemPileUuid);
@@ -916,6 +920,8 @@ export default class API {
 
         if (shouldBeDeleted) {
             await API._deleteItemPile(itemPileUuid);
+        }else if(lib.isItemPileEmpty(itemPileActor)){
+            await lib.clearItemPileSharingData(itemPileActor);
         }
 
         return transferData;
@@ -923,7 +929,6 @@ export default class API {
     }
 
     /* --- ITEM AND ATTRIBUTE METHODS --- */
-
 
     /**
      * Adds item to an actor, increasing item quantities if matches were found
@@ -936,7 +941,7 @@ export default class API {
      */
     static async addItems(target, items, { interactionId = false }={}) {
 
-        const hookResult = Hooks.call(HOOKS.ITEM.PRE_ADD, target, items);
+        const hookResult = Hooks.call(HOOKS.ITEM.PRE_ADD, target, items, interactionId);
         if (hookResult === false) return;
 
         const targetUuid = lib.getUuid(target);
@@ -1054,7 +1059,7 @@ export default class API {
      */
     static async removeItems(target, items, { interactionId = false }={}) {
 
-        const hookResult = Hooks.call(HOOKS.ITEM.PRE_REMOVE, target, items);
+        const hookResult = Hooks.call(HOOKS.ITEM.PRE_REMOVE, target, items, interactionId);
         if (hookResult === false) return;
 
         const targetUuid = lib.getUuid(target);
@@ -1183,7 +1188,7 @@ export default class API {
      */
     static async transferItems(source, target, items, { interactionId = false }={}) {
 
-        const hookResult = Hooks.call(HOOKS.ITEM.PRE_TRANSFER, source, target, items);
+        const hookResult = Hooks.call(HOOKS.ITEM.PRE_TRANSFER, source, target, items, interactionId);
         if (hookResult === false) return;
 
         const sourceUuid = lib.getUuid(source);
@@ -1240,8 +1245,6 @@ export default class API {
 
         if (!isEverything) {
 
-            await lib.setItemPileSharingData(sourceUuid, targetUuid, { items: itemsAdded });
-
             await itemPileSocket.executeForEveryone(SOCKET_HANDLERS.CALL_HOOK, HOOKS.ITEM.TRANSFER, sourceUuid, targetUuid, itemsAdded, userId, interactionId);
 
             const macroData = {
@@ -1259,8 +1262,14 @@ export default class API {
             await API.rerenderItemPileInventoryApplication(sourceUuid, shouldBeDeleted);
             await API.rerenderItemPileInventoryApplication(targetUuid);
 
+            const itemPile = await fromUuid(sourceUuid);
+
             if (shouldBeDeleted) {
                 await API._deleteItemPile(sourceUuid);
+            } else if (lib.isItemPileEmpty(itemPile)) {
+                await lib.clearItemPileSharingData(itemPile);
+            } else {
+                await lib.setItemPileSharingData(sourceUuid, targetUuid, { items: itemsAdded });
             }
 
         }
@@ -1281,7 +1290,7 @@ export default class API {
      */
     static async transferAllItems(source, target, { itemFilters = false, interactionId = false } = {}) {
 
-        const hookResult = Hooks.call(HOOKS.ITEM.PRE_TRANSFER_ALL, source, target, itemFilters);
+        const hookResult = Hooks.call(HOOKS.ITEM.PRE_TRANSFER_ALL, source, target, itemFilters, interactionId);
         if (hookResult === false) return;
 
         const sourceUuid = lib.getUuid(source);
@@ -1357,7 +1366,7 @@ export default class API {
      */
     static async addAttributes(target, attributes, { interactionId = false }={}) {
 
-        const hookResult = Hooks.call(HOOKS.ATTRIBUTE.PRE_ADD, target, attributes);
+        const hookResult = Hooks.call(HOOKS.ATTRIBUTE.PRE_ADD, target, attributes, interactionId);
         if (hookResult === false) return;
 
         const targetUuid = lib.getUuid(target);
@@ -1440,7 +1449,7 @@ export default class API {
      */
     static async removeAttributes(target, attributes, { interactionId = false }={}) {
 
-        const hookResult = Hooks.call(HOOKS.ATTRIBUTE.PRE_REMOVE, target, attributes);
+        const hookResult = Hooks.call(HOOKS.ATTRIBUTE.PRE_REMOVE, target, attributes, interactionId);
         if (hookResult === false) return;
 
         const targetUuid = lib.getUuid(target);
@@ -1548,7 +1557,7 @@ export default class API {
      */
     static async transferAttributes(source, target, attributes, { interactionId = false }={}) {
 
-        const hookResult = Hooks.call(HOOKS.ATTRIBUTE.PRE_TRANSFER, source, target, attributes);
+        const hookResult = Hooks.call(HOOKS.ATTRIBUTE.PRE_TRANSFER, source, target, attributes, interactionId);
         if (hookResult === false) return;
 
         const sourceUuid = lib.getUuid(source);
@@ -1611,8 +1620,6 @@ export default class API {
 
         if (!isEverything) {
 
-            await lib.setItemPileSharingData(sourceUuid, targetUuid, { currencies: attributesRemoved });
-
             await itemPileSocket.executeForEveryone(SOCKET_HANDLERS.CALL_HOOK, HOOKS.ATTRIBUTE.TRANSFER, sourceUuid, targetUuid, attributesRemoved, userId, interactionId);
 
             const macroData = {
@@ -1630,8 +1637,14 @@ export default class API {
             await API.rerenderItemPileInventoryApplication(sourceUuid, shouldBeDeleted);
             await API.rerenderItemPileInventoryApplication(targetUuid);
 
+            const itemPile = await fromUuid(sourceUuid)
+
             if (shouldBeDeleted) {
                 await API._deleteItemPile(sourceUuid);
+            } else if (lib.isItemPileEmpty(itemPile)) {
+                await lib.clearItemPileSharingData(itemPile);
+            } else {
+                await lib.setItemPileSharingData(sourceUuid, targetUuid, { currencies: attributesRemoved });
             }
 
         }
@@ -1649,9 +1662,9 @@ export default class API {
      *
      * @returns {Promise<object>}                       An object containing a key value pair of each attribute transferred, the key being the attribute path and its value being the quantity that was transferred
      */
-    static async transferAllAttributes(source, target,{ interactionId = false }={}) {
+    static async transferAllAttributes(source, target, { interactionId = false }={}) {
 
-        const hookResult = Hooks.call(HOOKS.ATTRIBUTE.PRE_TRANSFER_ALL, source, target);
+        const hookResult = Hooks.call(HOOKS.ATTRIBUTE.PRE_TRANSFER_ALL, source, target, interactionId);
         if (hookResult === false) return;
 
         const sourceUuid = lib.getUuid(source);
@@ -1734,7 +1747,7 @@ export default class API {
      */
     static async transferEverything(source, target, { itemFilters = false, interactionId = false } = {}) {
 
-        const hookResult = Hooks.call(HOOKS.PRE_TRANSFER_EVERYTHING, source, target, itemFilters);
+        const hookResult = Hooks.call(HOOKS.PRE_TRANSFER_EVERYTHING, source, target, itemFilters, interactionId);
         if (hookResult === false) return;
 
         const sourceUuid = lib.getUuid(source);
@@ -2040,7 +2053,7 @@ export default class API {
 
         } else {
 
-            const result = await DropItemDialog.query(itemData, droppableDocuments);
+            const result = await DropItemDialog.query(itemData, droppableDocuments[0]);
 
             if (!result) return;
             action = result.action;
@@ -2133,11 +2146,10 @@ export default class API {
      * @param {object} position
      * @param {string/boolean} [pileActorName=false]
      * @param {array/boolean} [items=false]
-     * @param {array/boolean} [currencies=false]
      * @returns {Promise<string>}
      * @private
      */
-    static async _createItemPile(sceneId, position, { pileActorName = false, items = false, currencies = false } = {}) {
+    static async _createItemPile(sceneId, position, { pileActorName = false, items = false } = {}) {
 
         let pileActor;
 
@@ -2193,22 +2205,18 @@ export default class API {
         if (!pileActor.data.token.actorLink) {
 
             items = items ? items.map(itemData => itemData.item ?? itemData) : [];
-            currencies = currencies || [];
 
             overrideData['actorData'] = {
-                items: items,
-                ...currencies
+                items: items
             }
 
-            const data = { data: pileData, items: items, currencies: currencies };
+            const data = { data: pileData, items: items };
 
             overrideData = foundry.utils.mergeObject(overrideData, {
-                "img": getItemPileTokenImage(pileActor, data),
-                "scale": getItemPileTokenScale(pileActor, data),
-                "name": getItemPileName(pileActor, data),
+                "img": lib.getItemPileTokenImage(pileActor, data),
+                "scale": lib.getItemPileTokenScale(pileActor, data),
+                "name": lib.getItemPileName(pileActor, data),
             });
-
-            console.log(overrideData);
 
         }
 
@@ -2227,8 +2235,6 @@ export default class API {
      */
     static async _itemPileClicked(pileDocument) {
 
-        //debugger;
-
         if(!lib.isValidItemPile(pileDocument)) return;
 
         const pileToken = pileDocument.object;
@@ -2244,20 +2250,23 @@ export default class API {
 
         const maxDistance = data.distance ? data.distance : Infinity;
 
-        let validTokens;
+        let validTokens = [];
 
         if(canvas.tokens.controlled.length > 0){
             validTokens = [...canvas.tokens.controlled];
-        }else{
-            validTokens = [...canvas.tokens.placeables];
-            if(_token){
+            validTokens = validTokens.filter(token => token.document !== pileDocument);
+        }
+
+        if(!validTokens.length && !game.user.isGM){
+            validTokens.push(...canvas.tokens.placeables);
+            if (_token) {
                 validTokens.unshift(_token);
             }
         }
 
-        validTokens = validTokens.filter(token => {
+        validTokens = validTokens.filter(token => token.owner && token.document !== pileDocument).filter(token => {
             return lib.tokens_close_enough(pileToken, token, maxDistance) || game.user.isGM;
-        }).filter(token => token.owner && token.document !== pileDocument);
+        });
 
         if (!validTokens.length && !game.user.isGM) {
             lib.custom_warning(game.i18n.localize("ITEM-PILES.Errors.PileTooFar"), true);