+++ scripts/libwrapper.js
@@ -1,12 +1,12 @@
 import API from "./api.js";
 import CONSTANTS from "./constants.js";
-import { hotkeyState } from "./hotkeys.js";
+import { hotkeyActionState } from "./hotkeys.js";
 import { ItemPileInventory } from "./formapplications/itemPileInventory.js";
 
 export function registerLibwrappers() {
 
     libWrapper.register(CONSTANTS.MODULE_NAME, 'Token.prototype._onClickLeft2', function (wrapped, ...args) {
-        if (API.isValidItemPile(this.document) && !hotkeyState.ctrlDown) {
+        if (API.isValidItemPile(this.document) && hotkeyActionState.openPileInventory) {
             return API._itemPileClicked(this.document);
         }
         return wrapped(...args);
@@ -19,7 +19,7 @@ export function registerLibwrappers() {
         const documentId = element.parentElement.dataset.documentId;
         const document = this.constructor.collection.get(documentId);
 
-        if (API.isValidItemPile(document) && !hotkeyState.ctrlDown) {
+        if (API.isValidItemPile(document) && hotkeyActionState.openPileInventory) {
             return ItemPileInventory.show(document, game.user.character);
         }