+++ scripts/module.js
@@ -1,6 +1,7 @@
 import CONSTANTS from "./constants.js";
 import HOOKS from "./hooks.js";
 
+import flagManager from "./flagManager.js";
 import chatHandler from "./chathandler.js";
 import API from "./api.js";
 import * as lib from "./lib/lib.js";
@@ -10,7 +11,6 @@ import { registerSettings, checkSystem, migrateSettings, registerHandlebarHelper
 import { registerSocket } from "./socket.js";
 import { registerLibwrappers } from "./libwrapper.js";
 import { registerHotkeysPre, registerHotkeysPost } from "./hotkeys.js";
-import flagManager from "./flagManager.js";
 import { getActorCurrencies, getActorItems } from "./lib/lib.js";
 
 Hooks.once("init", async () => {
@@ -21,6 +21,7 @@ Hooks.once("init", async () => {
 
     Hooks.once("socketlib.ready", registerSocket);
     Hooks.on("canvasReady", module._canvasReady);
+    Hooks.on("preCreateToken", module._preCreatePile);
     Hooks.on("createToken", module._createPile);
     Hooks.on("deleteToken", module._deletePile);
     Hooks.on("dropCanvasData", module._dropData);
@@ -139,6 +140,14 @@ const module = {
         const tokens = [...canvas.tokens.placeables].map(token => token.document);
         for (const doc of tokens) {
             await API._initializeItemPile(doc);
+        }
+    },
+
+    async _preCreatePile(document){
+        if(!document.isLinked){
+            document.data.update({
+                [`actorData.flags.${CONSTANTS.MODULE_NAME}.-=${CONSTANTS.SHARING_DATA}`]: null
+            });
         }
     },