+++ scripts/socket.js
@@ -3,7 +3,6 @@ import CONSTANTS from "./constants.js";
 import API from "./api.js";
 import { ItemPileInventory } from "./formapplications/itemPileInventory.js";
 import chatHandler from "./chathandler.js";
-import flagManager from "./flagManager.js";
 import { TradingHandler } from "./formapplications/tradingApp.js";
 
 export const SOCKET_HANDLERS = {
@@ -11,7 +10,6 @@ export const SOCKET_HANDLERS = {
      * Generic sockets
      */
     CALL_HOOK: "callHook",
-
 
     /**
      * Chat messages